# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelHub",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat/conversations": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Conversation Endpoint",
        "operationId": "create_conversation_endpoint_api_chat_conversations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Conversations Endpoint",
        "operationId": "list_conversations_endpoint_api_chat_conversations_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/conversations/{conv_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Conversation Endpoint",
        "operationId": "get_conversation_endpoint_api_chat_conversations__conv_id__get",
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete Conversation Endpoint",
        "operationId": "delete_conversation_endpoint_api_chat_conversations__conv_id__delete",
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Conversation Endpoint Api Chat Conversations  Conv Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/conversations/{conv_id}/messages": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Add Message Endpoint",
        "operationId": "add_message_endpoint_api_chat_conversations__conv_id__messages_post",
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conv Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Messages Endpoint",
        "operationId": "list_messages_endpoint_api_chat_conversations__conv_id__messages_get",
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conv Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/tools/{tool_name}": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Execute Tool Endpoint",
        "operationId": "execute_tool_endpoint_api_chat_tools__tool_name__post",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Arguments"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Execute Tool Endpoint Api Chat Tools  Tool Name  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/novel/{novel_id}/markdown": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Novel Markdown Endpoint",
        "operationId": "export_novel_markdown_endpoint_api_export_novel__novel_id__markdown_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/novel/{novel_id}/json": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Novel Json Endpoint",
        "operationId": "export_novel_json_endpoint_api_export_novel__novel_id__json_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/novel/{novel_id}/wiki/markdown": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Wiki Markdown Endpoint",
        "operationId": "export_wiki_markdown_endpoint_api_export_novel__novel_id__wiki_markdown_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationCreate": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": "New Chat"
          }
        },
        "type": "object",
        "required": [
          "novel_id"
        ],
        "title": "ConversationCreate"
      },
      "ConversationList": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationResponse"
            },
            "type": "array",
            "title": "Conversations"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "conversations",
          "total"
        ],
        "title": "ConversationList"
      },
      "ConversationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "default": "user"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "tool_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Calls",
            "default": []
          }
        },
        "type": "object",
        "title": "MessageCreate"
      },
      "MessageList": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "type": "array",
            "title": "Messages"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "messages",
          "total"
        ],
        "title": "MessageList"
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "conversation_id": {
            "type": "integer",
            "title": "Conversation Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          },
          "tool_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "role",
          "content",
          "citations",
          "tool_calls",
          "created_at"
        ],
        "title": "MessageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}